app/requirements.txt
flask==2.3.3  
spotipy==2.23.0  
python-dotenv==1.0.0  
requests==2.31.0